handle recent go versions
if I try the following on a recent go version (e.g. `go1.20.2`) I get:

```bash
go run main.go
# command-line-arguments
./main.go:80:19: undefined: NewProcess
```

this PR just moves the code from `process.go` into `main.go` to fix it so that `go run main.go` works
